Updated 8/2/2021

+++ EnGAGE/Scene.h
@@ -6,7 +6,7 @@ class Scene
 public:
 	virtual ~Scene() = default;
 
-	virtual void Update(double delta) noexcept {};
+	virtual void Update(float delta) noexcept {};
 	virtual void Render() noexcept {};
 	virtual void ImGui() noexcept {};
 };
